fixing network index relations

+++ clc/modules/cluster-manager/src/main/java/com/eucalyptus/cluster/VmInstances.java
@@ -66,7 +66,6 @@
 package com.eucalyptus.cluster;
 
 import java.security.MessageDigest;
-import java.util.Collection;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.zip.Adler32;
@@ -81,7 +80,6 @@ import com.eucalyptus.auth.Permissions;
 import com.eucalyptus.auth.policy.PolicySpec;
 import com.eucalyptus.auth.principal.Account;
 import com.eucalyptus.cloud.CloudMetadata.VmInstanceMetadata;
-import com.eucalyptus.cluster.VmInstance.Reason;
 import com.eucalyptus.cluster.VmInstance.VmState;
 import com.eucalyptus.cluster.callback.TerminateCallback;
 import com.eucalyptus.component.Dispatcher;
@@ -98,9 +96,6 @@ import com.eucalyptus.context.Contexts;
 import com.eucalyptus.crypto.Digest;
 import com.eucalyptus.entities.Entities;
 import com.eucalyptus.entities.TransactionException;
-import com.eucalyptus.entities.TransactionExecutionException;
-import com.eucalyptus.entities.TransientEntityException;
-import com.eucalyptus.network.NetworkGroup;
 import com.eucalyptus.network.NetworkGroups;
 import com.eucalyptus.records.EventRecord;
 import com.eucalyptus.records.EventType;
@@ -117,7 +112,6 @@ import com.eucalyptus.ws.client.ServiceDispatcher;
 import com.google.common.base.Function;
 import com.google.common.base.Functions;
 import com.google.common.base.Predicate;
-import com.google.common.collect.Collections2;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
 import edu.ucsb.eucalyptus.msgs.AttachedVolume;
@@ -133,7 +127,7 @@ public class VmInstances {
   @ConfigurableField( description = "Amount of time (in milliseconds) that a terminated VM will continue to be reported.", initial = "" + 60 * 60 * 1000 )
   public static Long         BURY_TIME                     = 60 * 60 * 1000l;
   @ConfigurableField( description = "Maximum amount of time (in seconds) that the network topology service takes to propagate state changes.", initial = "" + 60 * 60 * 1000 )
-  public static Integer      NETWORK_METADATA_REFRESH_TIME = 15;
+  public static Long         NETWORK_METADATA_REFRESH_TIME = 15l;
   @ConfigurableField( description = "Prefix to use for instance MAC addresses.", initial = "d0:0d" )
   public static String       MAC_PREFIX                    = "d0:0d";
   @ConfigurableField( description = "Subdomain to use for instance DNS.", initial = ".eucalyptus", changeListener = SubdomainListener.class )
